Refactor GeoJS extension code into separate geojsextension.ts file
